requirements.txt
streamlit
pandas
yfinance
python-dotenv
openai
pypdf
pydantic-ai
google-generativeai
html5lib
scikit-learn
pydantic-ai-slim[duckduckgo]
pydantic-ai-slim[tavily]
nest_asyncio